feat: Add option to disable processing the `format` keyword
Prior to draft versions 2019-09, format validation is enabled by default
(though should be able to be disabled), and vice versa afterword.
This patch adds `should_validate_formats` to CompilationOptions to force
the option one way or another.

If not specified, it will fallback to a default based on the draft
version (enabled by default before draft 2019-09).

Closes #261